Simplify the control code more
Instead of opening a websocket server, pass the control channel to the
WebViewPage through ::setWebChannel().

The JS side is able to access this transport as qt.webChannelTransport
without any socket connection.

The control channel is also limited to the main view for now, and some
object were renamed to reflect that.